triangle can be drawn now

+++ src/ven/render/ven_swap_chain.cpp
@@ -1,4 +1,5 @@
 #include "ven_swap_chain.h"
+#include <cstdint>
 
 ven::SwapChain::SwapChain(ven::Device& deviceRef, VkExtent2D extent)
     : device{deviceRef}
@@ -22,7 +23,7 @@ ven::SwapChain::~SwapChain() {
     swapChain = nullptr;
   }
 
-  for (int i = 0; i < (int)depthImages.size(); i++) {
+  for (uint32_t i = 0; i < depthImages.size(); i++) {
     vkDestroyImageView(device.device(), depthImageViews[i], nullptr);
     vkDestroyImage(device.device(), depthImages[i], nullptr);
     vkFreeMemory(device.device(), depthImageMemorys[i], nullptr);
@@ -288,7 +289,7 @@ void ven::SwapChain::createDepthResources() {
   depthImageMemorys.resize(imageCount());
   depthImageViews.resize(imageCount());
 
-  for (int i = 0; i < (int)depthImages.size(); i++) {
+  for (uint32_t i = 0; i < depthImages.size(); i++) {
     VkImageCreateInfo imageInfo{};
     imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
     imageInfo.imageType = VK_IMAGE_TYPE_2D;
